### 2024-12-10 #7555 @willu47
> 👋 @willu47 - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Yes. I can review this.
