### 2024-12-11 #7555 @jmlarson1
> 👋 @jmlarson1 - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I am unable to review this package at this time.
