+++ README.md
@@ -1,0 +1,104 @@
+# Repository Coverage
+
+[Full report](https://htmlpreview.github.io/?https://github.com/alan-turing-institute/autoemulate/blob/python-coverage-comment-action-data/htmlcov/index.html)
+
+| Name                                                                         |    Stmts |     Miss |   Cover |   Missing |
+|----------------------------------------------------------------------------- | -------: | -------: | ------: | --------: |
+| autoemulate/\_\_init\_\_.py                                                  |        0 |        0 |    100% |           |
+| autoemulate/compare.py                                                       |      169 |       22 |     87% |240-244, 278, 315, 317, 320, 338, 342, 345, 359, 465, 521-532, 594-603, 625 |
+| autoemulate/cross\_validate.py                                               |       44 |        7 |     84% |61-79, 150 |
+| autoemulate/data\_splitting.py                                               |        6 |        0 |    100% |           |
+| autoemulate/datasets.py                                                      |       12 |        0 |    100% |           |
+| autoemulate/emulators/\_\_init\_\_.py                                        |       27 |        0 |    100% |           |
+| autoemulate/emulators/conditional\_neural\_process.py                        |      102 |        3 |     97% |253-254, 295 |
+| autoemulate/emulators/conditional\_neural\_process\_attn.py                  |        9 |        0 |    100% |           |
+| autoemulate/emulators/gaussian\_process.py                                   |      104 |        5 |     95% |73, 259, 266, 282, 289 |
+| autoemulate/emulators/gaussian\_process\_mogp.py                             |       33 |       18 |     45% |18, 35-40, 59-65, 71-75, 79, 82 |
+| autoemulate/emulators/gaussian\_process\_mt.py                               |       97 |        7 |     93% |71, 240, 246, 265, 271, 286, 290 |
+| autoemulate/emulators/gaussian\_process\_sklearn.py                          |       42 |        0 |    100% |           |
+| autoemulate/emulators/gaussian\_process\_utils/\_\_init\_\_.py               |        3 |        0 |    100% |           |
+| autoemulate/emulators/gaussian\_process\_utils/early\_stopping\_criterion.py |       11 |        2 |     82% |    58, 63 |
+| autoemulate/emulators/gaussian\_process\_utils/poly\_mean.py                 |       23 |        0 |    100% |           |
+| autoemulate/emulators/gaussian\_process\_utils/polynomial\_features.py       |       35 |        0 |    100% |           |
+| autoemulate/emulators/gradient\_boosting.py                                  |       42 |        0 |    100% |           |
+| autoemulate/emulators/light\_gbm.py                                          |       52 |        0 |    100% |           |
+| autoemulate/emulators/neural\_net\_sk.py                                     |       42 |        0 |    100% |           |
+| autoemulate/emulators/neural\_networks/\_\_init\_\_.py                       |        0 |        0 |    100% |           |
+| autoemulate/emulators/neural\_networks/cnp\_module.py                        |       47 |        0 |    100% |           |
+| autoemulate/emulators/neural\_networks/cnp\_module\_attn.py                  |       50 |        0 |    100% |           |
+| autoemulate/emulators/neural\_networks/datasets.py                           |       49 |        1 |     98% |        11 |
+| autoemulate/emulators/neural\_networks/gp\_module.py                         |       23 |        0 |    100% |           |
+| autoemulate/emulators/neural\_networks/losses.py                             |       12 |        0 |    100% |           |
+| autoemulate/emulators/polynomials.py                                         |       33 |        0 |    100% |           |
+| autoemulate/emulators/radial\_basis\_functions.py                            |       34 |        0 |    100% |           |
[…]
