### 2024-12-18 d3e9a76
fix scikit-learn dependency error
