+++ pyproject.toml
@@ -17,7 +17,7 @@ include = ["misc/AE_logo_final.png"]
 python = ">=3.10,<3.13"
 mogp-emulator = "^0.7.2"
 matplotlib = "^3.7.2"
-scikit-learn = "^1.3.0"
+scikit-learn = ">=1.3.0,<1.6.0"
 pandas = "^2.1"
 torch = "^2.1.0"
 skorch = "^0.15.0"
