+++ _sources/getting-started/index.md
@@ -1,0 +1,5 @@
+# Getting started
+
+`AutoEmulate` is simple to use out of the box, but there are lots of options for
+more fine grained control. Here, we'll show you how to install and use `AutoEmulate`
+and how to contribute to the project.
