+++ _sources/reference/emulators/light_gbm.rst
@@ -1,0 +1,6 @@
+autoemulate.emulators.light_gbm
+===============================
+
+.. automodule:: autoemulate.emulators.light_gbm
+   :members:
+   :show-inheritance:
