+++ _sources/reference/emulators/gaussian_process_sklearn.rst
@@ -1,0 +1,6 @@
+autoemulate.emulators.gaussian_process
+======================================
+
+.. automodule:: autoemulate.emulators.gaussian_process_sklearn
+   :members:
+   :show-inheritance:
