+++ _sources/reference/emulators/conditional_neural_process.rst
@@ -1,0 +1,6 @@
+autoemulate.emulators.conditional_neural_process
+================================================
+
+.. automodule:: autoemulate.emulators.conditional_neural_process
+    :members:
+    :show-inheritance:
