### 2025-01-03 #7626 @editorialbot
[REVIEW]: AutoEmulate: A Python package for semi-automated emulation
**Submitting author:** @mastoffel (<a href="http://orcid.org/0000-0003-4030-3543">Martin Stoffel</a>)
**Repository:** https://github.com/alan-turing-institute/autoemulate
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.1
