### 2025-01-24 #7626 @willu47
@mastoffel - could you clarify the contribution of the two non-contributing paper authors?
