### 2025-01-24 #7626 @mastoffel
@willu47 if you mean non-contributing to the code, that's Eric and Steve. They came up with the idea for the project and contributed ideas / project management since. Should I create a contribution section in the paper or is this sufficient here?
