### 2025-01-27 alan-turing-institute/autoemulate#284 @willu47
JOSS REVIEW: Consider adding a credit attribution statement to the paper
Part of [JOSS REVIEW](https://github.com/openjournals/joss-reviews/issues/7626)

Consider including a CREDIT attribution statement in your paper. This is not a formal requirement of JOSS (but perhaps it should be), but it would clarify the contribution of all authors, given that not all are included in the contributor list on GitHub.
