### 2025-01-27 #7626 @willu47
@crvernon - I don't believe this is a COI, but in the interests of transparency, `autoemulate` depends upon a Python library I maintain `SALib`.
