### 2025-01-27 alan-turing-institute/autoemulate#285 @willu47
JOSS REVIEW: Installation
Part of [JOSS REVIEW](https://github.com/openjournals/joss-reviews/issues/7626#)

# Install from pip

- [x] Add compatible Python version to the installation docs

Steps I followed:

- Created a new conda environment using mamba, defaulted to Python=3.13
- `pip install autoemulate` - FAILED (<3.13 constraint on pip)
- `mamba install python=3.12`
- `pip install autoemulate` - PASSED

I ran the quickstart and received an error due to sci-kit learn dependency.

- [x] Release a patch for v2 on pypi to fix this bug and incorporate the pin of sci-kit learn <1.6.0

Ultimately, you'll want to fix this issue.

After installing scikit-learn 1.5.2, I was able to run the quick-start successfully

# Editable install from local repo/Github clone

1. Git clone to local folder
2. `pip install -e .` into my conda environment
3. `mamba install pytest`
4. `pytest` # ran the tests locally and all passed.

- [x] Review the warnings in the test warnings summary

<details><summary>Warnings Summary</summary>

```python

tests/models/test_gptorch.py: 5 warnings
tests/test_estimators.py: 30 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/gpytorch/models/exact_gp.py:284: GPInputWarning: The input matches the stored training data. Did you forget to call model.train()?
    warnings.warn(

tests/test_end_to_end.py: 16 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/metrics/_regression.py:1211: UndefinedMetricWarning: R^2 score is not well-defined with less than two samples.
    warnings.warn(msg, UndefinedMetricWarning)

tests/test_end_to_end.py::test_run_param_search
tests/test_ui.py::test_param_search
tests/test_ui.py::test_save_load_with_param_search
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/model_selection/_search.py:320: UserWarning: The total space of parameters 1 is smaller than n_iter=2. Running 1 iterations. For exhaustive searches, use GridSearchCV.
    warnings.warn(

tests/test_end_to_end.py::test_run_parallel
tests/test_end_to_end.py::test_run_parallel
tests/test_end_to_end.py::test_run_parallel
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/joblib/externals/loky/backend/fork_exec.py:38: DeprecationWarning: This process (pid=85038) is multi-threaded, use of fork() may lead to deadlocks in the child.
    pid = os.fork()

tests/test_end_to_end.py::test_run_parallel
tests/test_end_to_end.py::test_run_parallel
tests/test_ui.py::test_save_load_with_param_search
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/skorch/net.py:2231: FutureWarning: You are using `torch.load` with `weights_only=False` (the current default value), which uses the default pickle module implicitly. It is possible to construct malicious pickle data which will execute arbitrary code during unpickling (See https://github.com/pytorch/pytorch/blob/main/SECURITY.md#untrusted-models for more details). In a future release, the default value for `weights_only` will be flipped to `True`. This limits the functions that could be executed during unpickling. Arbitrary objects will no longer be allowed to be loaded via this mode unless they are explicitly allowlisted by the user via `torch.serialization.add_safe_globals`. We recommend you start setting `weights_only=True` for any use case where you don't have full control of the loaded file. Please open an issue on GitHub for any issues related to this experimental feature.
    cuda_attrs = torch.load(f, **load_kwargs)

tests/test_sensitivity_analysis.py::test_sobol_analysis
tests/test_sensitivity_analysis.py::test_sobol_analysis_2d
tests/test_sensitivity_analysis.py::test_sobol_results_to_df
tests/test_sensitivity_analysis.py::test_validate_input
tests/test_sensitivity_analysis.py::test_validate_input_valid
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/SALib/util/__init__.py:274: FutureWarning: unique with argument that is not not a Series, Index, ExtensionArray, or np.ndarray is deprecated and will raise in a future version.
    names = list(pd.unique(groups))

tests/test_ui.py: 16 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/base.py:1473: DataConversionWarning: A column-vector y was passed when a 1d array was expected. Please change the shape of y to (n_samples,), for example using ravel().
    return fit_method(estimator, *args, **kwargs)

tests/test_ui.py: 48 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/utils/validation.py:1339: DataConversionWarning: A column-vector y was passed when a 1d array was expected. Please change the shape of y to (n_samples, ), for example using ravel().
    y = column_or_1d(y, warn=True)

tests/test_ui.py: 16 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/neural_network/_multilayer_perceptron.py:1631: DataConversionWarning: A column-vector y was passed when a 1d array was expected. Please change the shape of y to (n_samples, ), for example using ravel().
    y = column_or_1d(y, warn=True)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
```</summary>
```python
tests/models/test_gptorch.py: 5 warnings
tests/test_estimators.py: 30 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/gpytorch/models/exact_gp.py:284: GPInputWarning: The input matches the stored training data. Did you forget to call model.train()?
    warnings.warn(

tests/test_end_to_end.py: 16 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/metrics/_regression.py:1211: UndefinedMetricWarning: R^2 score is not well-defined with less than two samples.
    warnings.warn(msg, UndefinedMetricWarning)

tests/test_end_to_end.py::test_run_param_search
tests/test_ui.py::test_param_search
tests/test_ui.py::test_save_load_with_param_search
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/model_selection/_search.py:320: UserWarning: The total space of parameters 1 is smaller than n_iter=2. Running 1 iterations. For exhaustive searches, use GridSearchCV.
    warnings.warn(

tests/test_end_to_end.py::test_run_parallel
tests/test_end_to_end.py::test_run_parallel
tests/test_end_to_end.py::test_run_parallel
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/joblib/externals/loky/backend/fork_exec.py:38: DeprecationWarning: This process (pid=85038) is multi-threaded, use of fork() may lead to deadlocks in the child.
    pid = os.fork()

tests/test_end_to_end.py::test_run_parallel
tests/test_end_to_end.py::test_run_parallel
tests/test_ui.py::test_save_load_with_param_search
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/skorch/net.py:2231: FutureWarning: You are using `torch.load` with `weights_only=False` (the current default value), which uses the default pickle module implicitly. It is possible to construct malicious pickle data which will execute arbitrary code during unpickling (See https://github.com/pytorch/pytorch/blob/main/SECURITY.md#untrusted-models for more details). In a future release, the default value for `weights_only` will be flipped to `True`. This limits the functions that could be executed during unpickling. Arbitrary objects will no longer be allowed to be loaded via this mode unless they are explicitly allowlisted by the user via `torch.serialization.add_safe_globals`. We recommend you start setting `weights_only=True` for any use case where you don't have full control of the loaded file. Please open an issue on GitHub for any issues related to this experimental feature.
    cuda_attrs = torch.load(f, **load_kwargs)

tests/test_sensitivity_analysis.py::test_sobol_analysis
tests/test_sensitivity_analysis.py::test_sobol_analysis_2d
tests/test_sensitivity_analysis.py::test_sobol_results_to_df
tests/test_sensitivity_analysis.py::test_validate_input
tests/test_sensitivity_analysis.py::test_validate_input_valid
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/SALib/util/__init__.py:274: FutureWarning: unique with argument that is not not a Series, Index, ExtensionArray, or np.ndarray is deprecated and will raise in a future version.
    names = list(pd.unique(groups))

tests/test_ui.py: 16 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/base.py:1473: DataConversionWarning: A column-vector y was passed when a 1d array was expected. Please change the shape of y to (n_samples,), for example using ravel().
    return fit_method(estimator, *args, **kwargs)

tests/test_ui.py: 48 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/utils/validation.py:1339: DataConversionWarning: A column-vector y was passed when a 1d array was expected. Please change the shape of y to (n_samples, ), for example using ravel().
    y = column_or_1d(y, warn=True)

tests/test_ui.py: 16 warnings
  /Users/wusher_1/miniforge3/envs/auto-review/lib/python3.12/site-packages/sklearn/neural_network/_multilayer_perceptron.py:1631: DataConversionWarning: A column-vector y was passed when a 1d array was expected. Please change the shape of y to (n_samples, ), for example using ravel().
    y = column_or_1d(y, warn=True)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
```
