### 2025-01-27 alan-turing-institute/autoemulate#286 @willu47
JOSS REVIEW: Functionality
Part of [JOSS REVIEW](https://github.com/openjournals/joss-reviews/issues/7626#)

I have read through the quick starts and tutorials, and run the code.  A few small observations in the docs:

- [x] [Save the model step under quickstart](https://alan-turing-institute.github.io/autoemulate/getting-started/quickstart.html#saving-loading-models) has an incorrect model name `best_emulator` instead of `gp_final`
- [x] Typo [here](https://alan-turing-institute.github.io/autoemulate/tutorials/01_start.html#dimension-reduction) `be default` should be `by default`
- [x] Missing first "time taken" result [here](https://alan-turing-institute.github.io/autoemulate/tutorials/02_speed.html#minimising-computation-time)
