### 2025-01-27 alan-turing-institute/autoemulate#287 @willu47
JOSS REVIEW: Community guidelines
Part of [JOSS REVIEW](https://github.com/openjournals/joss-reviews/issues/7626#)

- [x] There are broken links in the `CONTRIBUTING.md` file.  Neither of the links in the [how to contribute](https://github.com/alan-turing-institute/autoemulate/blob/main/CONTRIBUTING.md#how-to-contribute) section work (these sections work in the documentation, I guess because the file is inserted into the relevant section).
- [x] There are no instructions on how to make a release or to deploy the package to PyPi (essential docs for maintainers)
- [x] Coupled with the above, I strongly encourage setting up a continuous deployment pipeline for deploying the package to PyPi pushing a tag (or making a release on Github), which greatly eases the process!
