### 2025-01-29 c4b4992
fix sensitivity test warnings
