### 2025-01-30 a938b66
update installation instructions with correct python versions
