### 2025-01-30 5f2908b
update tensor to not share memory to avoid warning
