### 2025-01-30 9ee857e
restrict compatible PyTorch versions
add safe globals to fix pytorch 2.6 error

modify safe globals

remove safe globals

add safe_globals

remove safe globals
