+++ pyproject.toml
@@ -19,7 +19,7 @@ mogp-emulator = "^0.7.2"
 matplotlib = "^3.7.2"
 scikit-learn = ">=1.3.0,<1.6.0"
 pandas = "^2.1"
-torch = ">=2.1.0"
+torch = "^2.1.0"
 skorch = "^0.15.0"
 scipy = "^1.11.3"
 numpy = "^1.24"
