### 2025-01-30 alan-turing-institute/autoemulate#285 @mastoffel
Thanks @willu47! All warnings have been fixed and Python version has been added to the installation docs and README in #288.
