+++ README.md
@@ -12,9 +12,9 @@
 | autoemulate/emulators/\_\_init\_\_.py                                        |       27 |        0 |    100% |           |
 | autoemulate/emulators/conditional\_neural\_process.py                        |      102 |        3 |     97% |253-254, 295 |
 | autoemulate/emulators/conditional\_neural\_process\_attn.py                  |        9 |        0 |    100% |           |
-| autoemulate/emulators/gaussian\_process.py                                   |      104 |        5 |     95% |73, 266, 289, 321, 332 |
+| autoemulate/emulators/gaussian\_process.py                                   |      104 |        7 |     93% |73, 266, 274, 289, 300, 321, 336 |
 | autoemulate/emulators/gaussian\_process\_mogp.py                             |       33 |       18 |     45% |18, 35-40, 59-65, 71-75, 79, 82 |
-| autoemulate/emulators/gaussian\_process\_mt.py                               |       97 |        7 |     93% |71, 240, 246, 265, 271, 286, 290 |
+| autoemulate/emulators/gaussian\_process\_mt.py                               |       97 |        8 |     92% |71, 240, 250, 254, 259, 265, 278, 286 |
 | autoemulate/emulators/gaussian\_process\_sklearn.py                          |       42 |        0 |    100% |           |
 | autoemulate/emulators/gaussian\_process\_utils/\_\_init\_\_.py               |        3 |        0 |    100% |           |
 | autoemulate/emulators/gaussian\_process\_utils/early\_stopping\_criterion.py |       11 |        2 |     82% |    58, 63 |
@@ -51,13 +51,13 @@
 | tests/models/test\_attn\_cnp.py                                              |      147 |        0 |    100% |           |
 | tests/models/test\_cnp.py                                                    |      108 |        0 |    100% |           |
 | tests/models/test\_cnp\_dataset.py                                           |       74 |        0 |    100% |           |
-| tests/models/test\_gptorch.py                                                |       66 |        4 |     94% |     62-65 |
+| tests/models/test\_gptorch.py                                                |       81 |        5 |     94% |     77-81 |
 | tests/test\_compare.py                                                       |      149 |        0 |    100% |           |
 | tests/test\_cross\_validate.py                                               |       97 |        3 |     97% | 56-59, 64 |
 | tests/test\_data\_splitting.py                                               |       11 |        0 |    100% |           |
 | tests/test\_datasets.py                                                      |       13 |        0 |    100% |           |
 | tests/test\_end\_to\_end.py                                                  |       39 |        0 |    100% |           |
-| tests/test\_estimators.py                                                    |       25 |        0 |    100% |           |
+| tests/test\_estimators.py                                                    |       26 |        0 |    100% |           |
 | tests/test\_experimental\_design.py                                          |       22 |        0 |    100% |           |
 | tests/test\_gaussian\_process\_utils.py                                      |       76 |        0 |    100% |           |
 | tests/test\_hyperparam\_searching.py                                         |       48 |        0 |    100% |           |
@@ -67,10 +67,10 @@
 | tests/test\_plotting.py                                                      |      194 |        7 |     96% |44, 54, 76-77, 85-86, 94 |
 | tests/test\_printing.py                                                      |       19 |        0 |    100% |           |
 | tests/test\_save.py                                                          |       62 |        2 |     97% |    30, 35 |
-| tests/test\_sensitivity\_analysis.py                                         |      111 |        0 |    100% |           |
-| tests/test\_ui.py                                                            |       57 |        0 |    100% |           |
+| tests/test\_sensitivity\_analysis.py                                         |      116 |        0 |    100% |           |
+| tests/test\_ui.py                                                            |       55 |        0 |    100% |           |
[…]
