### 2025-01-30 alan-turing-institute/autoemulate#285 @mastoffel
Released a new [pypi version](https://pypi.org/project/autoemulate/) fixing the sci-kit learn dependency issue.
