+++ tests/test_estimators.py
@@ -44,6 +44,8 @@ from autoemulate.emulators import SupportVectorMachines
         GaussianProcess(random_state=42),
     ]
 )
-@pytest.mark.filterwarnings("ignore::gpytorch.models.exact_gp.GPInputWarning")
+@pytest.mark.filterwarnings(
+    "ignore::gpytorch.models.exact_gp.GPInputWarning", "ignore::UserWarning"
+)
 def test_check_estimator(estimator, check):
     check(estimator)
