+++ autoemulate/metrics.py
@@ -2,7 +2,6 @@ import numpy as np
 from sklearn.metrics import r2_score
 from sklearn.metrics import root_mean_squared_error
 
-
 def rmse(y_true, y_pred, multioutput="uniform_average"):
     """Returns the root mean squared error.
 
@@ -38,3 +37,38 @@ METRIC_REGISTRY = {
     "rmse": rmse,
     "r2": r2,
 }
+
+
+def history_matching(obs, expectations, threshold=3.0, discrepancy=0.0, rank=1):
+    """
+    Perform history matching to compute implausibility and identify NROY and RO points.
+
+    Parameters:
+        obs (tuple): Observations as (mean, variance).
+        expectations (tuple): Predicted (mean, variance).
+        threshold (float): Implausibility threshold for NROY classification.
+        discrepancy (float or ndarray): Discrepancy value(s).
+        rank (int): Rank for implausibility calculation.
+
+    Returns:
+        dict: Contains implausibility (I), NROY indices, and RO indices.
+    """
+    obs_mean, obs_var = np.atleast_1d(obs[0]), np.atleast_1d(obs[1])
+    pred_mean, pred_var = np.atleast_1d(expectations[0]), np.atleast_1d(expectations[1])
+
+    discrepancy = np.atleast_1d(discrepancy)
+    n_obs = len(obs_mean)
+    rank = min(max(rank, 0), n_obs - 1)
+    #  Vs represents the total variance associated with the observations, predictions, and potential discrepancies. 
+    Vs = pred_var + discrepancy[:, np.newaxis] + obs_var[:, np.newaxis]
+    I = np.abs(obs_mean[:, np.newaxis] - pred_mean) / np.sqrt(Vs)
+    I_ranked = np.partition(I, rank, axis=0)[rank]
+
+    NROY = np.where(I_ranked <= threshold)[0]
+    RO = np.where(I_ranked > threshold)[0]
+
+    return {
+        "I": I_ranked,
+        "NROY": list(NROY),
+        "RO": list(RO)
+    }
