### 2025-02-03 0420064
Added MLE to the metrics and the same analysis to the notebook. It can benefit from removing some of teh hardcoded parameters and making them input.
