### 2025-02-07 f821e6b
add version numbering guide to release.md
