### 2025-02-07 alan-turing-institute/autoemulate#287 @mastoffel
The links are fixed, there is now a [release section](https://github.com/alan-turing-institute/autoemulate/blob/main/docs/community/release.md) in the docs, and we've added a [gh-actions workflow](https://github.com/alan-turing-institute/autoemulate/blob/main/.github/workflows/release.yaml) to automate the release upon pushing a tag, thanks for the suggestion @willu47!
