### 2025-02-07 c0c85aa
add pytorch extractor+tests
