### 2025-02-07 4ea7446
fix pytorch extractor tests
