### 2025-02-11 c456c38
add warning that scalers are removed
