### 2025-02-12 ce18abe
refactor extract pytorch model
