### 2025-02-12 40dbed2
return pytorch model in eval mode
