### 2025-02-13 b20e62f
adding select_kernel funciton to utils. This should be discussed, what is the best way of extracting the kernel
