+++ docs/_config.yml
@@ -1,6 +1,10 @@
 title: "AutoEmulate: A package for semi-automated emulation"
 author: Martin Stoffel
-# logo: logo.png
+logo: ../misc/AE_logo_final.png
+theme: 
+  primary_color: "#0077CC"  # A professional blue color
+  logo_only: true
+  display_navbar_logo_on_scroll: true
 
 # Force re-execution of notebooks on each build.
 # See https://jupyterbook.org/content/execute.html
