### 2025-02-17 6da5972
Adjust the tests to allow new addition to arguments in sensitivity analysis
