### 2025-02-18 5cc9112
Update tests/test_sensitivity_analysis.py
Co-authored-by: Sam Greenbury <50113363+sgreenbury@users.noreply.github.com>
