### 2025-02-18 #7626 @lbl59
@mastoffel Would it be possible to add `plt.show()` lines to the Quickstart example provided in the GitHub repo so beginner users can view the plots immediately in their window? The plots do not appear when running this example. For reference I am running Python 3.12.7
