### 2025-02-18 4aef268
parameter name change, checking for the length of obs and predictions to be matching
