### 2025-02-18 6ec96a7
fixing tests to accept new variable and type check
