+++ docs/_config.yml
@@ -33,6 +33,7 @@ repository:
 html:
   use_issues_button: true
   use_repository_button: true
+  default_mode: dark
 
 sphinx:
   extra_extensions:
