### 2025-02-21 7702508
add plt.show for terminal
