### 2025-02-21 2878310
refactor _display_plot
