### 2025-02-21 4a3894d
make sure plot_eval also shows in the terminal
