### 2025-02-21 bcda9a0
add calculate subplot function
