### 2025-02-21 a8de615
fix plotting tests
