### 2025-02-21 32fd5fe
show plots for sensitivity analysis
