### 2025-02-21 #7626 @mastoffel
thanks @lbl59 ! There's now a [PR](https://github.com/alan-turing-institute/autoemulate/pull/310) fixing this. We now check whether plots are run from a terminal and then use `plt.show` to show them.
