### 2025-02-25 0d5b06d
Update autoemulate/plotting.py
Co-authored-by: Radka Jersakova <r.jersakova@gmail.com>
