+++ _sources/reference/emulators/rbf.rst
@@ -1,0 +1,6 @@
+autoemulate.emulators.radial_basis_functions
+============================================
+
+.. automodule:: autoemulate.emulators.radial_basis_functions
+   :members:
+   :show-inheritance:
