+++ _sources/getting-started/index.md
@@ -1,0 +1,3 @@
+# Getting started
+
+Here, we'll show you how to install `AutoEmulate` and get started with the basics.
