### 2025-03-04 f9e7c5f
Added skeleton for active learning class
