### 2025-03-06 cd596b7
update contributing documentation to include local testing instructions
