### 2025-03-06 9776a43
move homepage content to getting started
