### 2025-03-06 0b3cf27
remove side bar from home page
