+++ docs/index.md
@@ -1,3 +1,7 @@
+```{raw} html
+<body class="home-page">
+```
+
 # Welcome to AutoEmulate
 
 `AutoEmulate` is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration.
