+++ docs/_config.yml
@@ -42,7 +42,10 @@ sphinx:
     - 'sphinx.ext.autodoc'
     - 'sphinx.ext.napoleon'
     - 'sphinx.ext.viewcode'
+  extra_css: ["_static/custom.css"]
   config:
+    html_context:
+      body_classes: home-page
     add_module_names: False
     autodoc_typehints: none
     autodoc_member_order: 'bysource'
