+++ docs/index.md
@@ -2,7 +2,7 @@
 <body class="home-page">
 ```
 
-# Welcome to AutoEmulate
+# Emulators for complex simulations
 
 `AutoEmulate` is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration.
 
