+++ docs/index.md
@@ -4,8 +4,6 @@
 
 # Emulators for complex simulations
 
-`AutoEmulate` is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration.
-
 ```{button-ref} getting-started/index
 :ref-type: doc
 :color: primary
@@ -14,6 +12,8 @@
 
 🚀 Get started
 ```
+
+`AutoEmulate` is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration.
 
 ::::{grid} 1 1 2 3
 :gutter: 2
