+++ docs/index.md
@@ -13,7 +13,11 @@
 🚀 Get started
 ```
 
-`AutoEmulate` is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration.
+```{raw} html
+<p class="explainer-text">
+    AutoEmulate is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration
+</p>
+```
 
 ::::{grid} 1 1 2 3
 :gutter: 2
@@ -48,3 +52,7 @@ Report bugs or request new features
 The AutoEmulate API
 :::
 ::::
+
+```{raw} html
+</body>
+```
