+++ docs/index.md
@@ -15,7 +15,7 @@
 
 ```{raw} html
 <p class="explainer-text">
-    AutoEmulate is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration
+    <i>AutoEmulate</i> is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration
 </p>
 ```
 
