### 2025-03-07 6e5b4f6
What are we trying to solve
