+++ docs/index.md
@@ -18,12 +18,21 @@
     <i>AutoEmulate</i> is a Python library that makes it easy to create accurate and efficient emulators for complex simulations. Under the hood, the package runs a complete machine learning pipeline to compare and optimise a wide range of models, and provides functions for downstream tasks like prediction, sensitivity analysis and calibration.
 </p><br><br>
 <div style="display: flex; align-items: center; justify-content: space-between; width: 100%;">
-    <p class="left-text" style="flex: 1;">
+    <p class="left-text">
         <b>What are we trying to solve?</b><br>
         For complex systems, the computational cost of simulation poses a significant bottleneck.
         As an example, Cardio-vascular CFD models can take hours or even days to run.
     </p>
     <img style="flex: 1; max-width: 50%; text-align: right;" src="./_static/heart.png" alt="Heart Image" /><br>
+</div>
+<div style="display: flex; align-items: center; justify-content: space-between; width: 100%;">
+    <img style="flex: 1; max-width: 50%; text-align: left;" src="./_static/logo.png" alt="Logo" />
+    <p class="right-text">
+        <b>Introducing <i>AutoEmulate</i></b><br>
+        Replace complex simulations with fast, accurate emulators.
+        Automatically fit and compare various emulators, ranging from simple models like Radial Basis Functions and Second Order Polynomials to more complex models like Support Vector Machines, Gaussian Processes and Conditional Neural Processes.
+        
+    </p><br><br>
 </div>
 ```
 
