+++ autoemulate/preprocess_target.py
@@ -1,0 +1,447 @@
+import numpy as np
+from sklearn.base import BaseEstimator
+from sklearn.base import TransformerMixin
+from sklearn.decomposition import PCA
+from sklearn.preprocessing import StandardScaler
+from sklearn.utils.metaestimators import _BaseComposition
+
+
+class OutputOnlyPreprocessor(BaseEstimator, TransformerMixin):
+    """
+    A custom transformer for sklearn Pipeline that applies preprocessing
+    to only the y (output) values during model training.
+
+    This preprocessor can be used to apply dimensionality reduction (like PCA)
+    or other transformations to target variables before model training.
+
+    Parameters
+    ----------
+    methods : str or list of str, default=None
+        The preprocessing method(s) to apply to y.
+        Options include:
+        - 'pca': Apply PCA to reduce dimensionality
+        - 'standardize': Apply StandardScaler
+        - custom methods can be added
+
+    n_components : int or float, default=None
+        For PCA, the number of components to keep.
+        If None and method includes 'pca', keeps all components.
+
+    Attributes
+    ----------
+    transformers_ : dict
+        Dictionary of fitted transformer objects with method names as keys.
+    inverse_transformers_ : dict
+        Dictionary of inverse transformation methods for each transformer.
+    """
+
+    def __init__(self, methods=None, n_components=None):
+        self.methods = methods
+        self.n_components = n_components
+        self.transformers_ = {}
+        self.inverse_transformers_ = {}
+
+    def fit(self, X, y=None):
+        """
+        Fit all the transformers on the output data.
+
+        Parameters
+        ----------
+        X : array-like
+            Input features (not used in this transformer)
+        y : array-like
+            Target values to be preprocessed
+
+        Returns
+        -------
+        self : object
+            Returns self.
+        """
+        if y is None:
+            return self
+
+        # Ensure y is a numpy array
+        y = np.asarray(y)
+
+        # Initialize transformers based on methods
+        if self.methods is None:
+            return self
+
+        methods = self.methods if isinstance(self.methods, list) else [self.methods]
+
+        for method in methods:
+            if method == "pca":
+                transformer = PCA(n_components=self.n_components)
+                self.transformers_[method] = transformer.fit(y)
+                self.inverse_transformers_[
+                    method
+                ] = lambda y_transformed, transformer=transformer: transformer.inverse_transform(
+                    y_transformed
+                )
+
+            elif method == "standardize":
+                transformer = StandardScaler()
+                self.transformers_[method] = transformer.fit(y)
+                self.inverse_transformers_[
+                    method
+                ] = lambda y_transformed, transformer=transformer: transformer.inverse_transform(
+                    y_transformed
+                )
+
+        return self
+
+    def transform(self, X, y=None):
+        """
+        Transform input data X (pass-through) and output data y (if provided).
+        During fitting in a pipeline, y is provided and transformed.
+
+        Parameters
+        ----------
+        X : array-like
+            Input features (returned unchanged)
+        y : array-like, default=None
+            Target values to transform
+
+        Returns
+        -------
+        X : array-like
+            Original input features
+        y_transformed : array-like or None
+            Transformed target values (if y was provided)
+        """
+        # During pipeline.fit(), return X unchanged and transformed y
+        if y is not None and self.methods is not None:
+            y = np.asarray(y)
+            methods = self.methods if isinstance(self.methods, list) else [self.methods]
+
+            y_transformed = y.copy()
+            for method in methods:
+                if method in self.transformers_:
+                    y_transformed = self.transformers_[method].transform(y_transformed)
+
[…]
