### 2025-03-07 25dfb88
Added a VAE class that returns the latent instead of Y, this class has a inverse_transform that runs the decoder
