+++ autoemulate/compare.py
@@ -475,6 +475,7 @@ class AutoEmulate:
                 methods=self.preprocess_outputs
             ).inverse_transform(y_pred)
 
+
         y_true = self.y[self.test_idxs]
 
         scores = {}
