+++ autoemulate/preprocess_target.py
@@ -1,19 +1,333 @@
 import numpy as np
-from sklearn.base import BaseEstimator
-from sklearn.base import TransformerMixin
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import StandardScaler
-from sklearn.utils.metaestimators import _BaseComposition
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+from torch.utils.data import DataLoader, TensorDataset
+from sklearn.base import BaseEstimator, TransformerMixin
+
+class VAE(nn.Module):
+    """
+    Variational Autoencoder implementation in PyTorch.
+    
+    Parameters
+    ----------
+    input_dim : int
+        Dimensionality of input data.
+    hidden_dims : list
+        List of hidden dimensions for encoder and decoder networks.
+    latent_dim : int
+        Dimensionality of the latent space.
+    """
+    def __init__(self, input_dim, hidden_dims, latent_dim):
+        super(VAE, self).__init__()
+        
+        # Encoder layers
+        encoder_layers = []
+        prev_dim = input_dim
+        for dim in hidden_dims:
+            encoder_layers.append(nn.Linear(prev_dim, dim))
+            encoder_layers.append(nn.ReLU())
+            prev_dim = dim
+        
+        self.encoder = nn.Sequential(*encoder_layers)
+        self.fc_mu = nn.Linear(hidden_dims[-1], latent_dim)
+        self.fc_var = nn.Linear(hidden_dims[-1], latent_dim)
+        
+        # Decoder layers
+        decoder_layers = []
+        prev_dim = latent_dim
+        for dim in reversed(hidden_dims):
+            decoder_layers.append(nn.Linear(prev_dim, dim))
+            decoder_layers.append(nn.ReLU())
+            prev_dim = dim
+        decoder_layers.append(nn.Linear(prev_dim, input_dim))
+        
+        self.decoder = nn.Sequential(*decoder_layers)
+    
+    def encode(self, x):
+        """Encode input to mean and log variance of latent distribution."""
+        x = self.encoder(x)
+        mu = self.fc_mu(x)
+        log_var = self.fc_var(x)
+        return mu, log_var
+    
+    def reparameterize(self, mu, log_var):
+        """Sample from latent distribution using reparameterization trick."""
+        std = torch.exp(0.5 * log_var)
+        eps = torch.randn_like(std)
+        z = mu + eps * std
+        return z
+    
+    def decode(self, z):
+        """Decode latent representation back to original space."""
+        return self.decoder(z)
+    
+    def forward(self, x):
+        """Full forward pass through the VAE."""
+        mu, log_var = self.encode(x)
+        z = self.reparameterize(mu, log_var)
+        recon = self.decode(z)
+        return recon, mu, log_var
+
+    def loss_function(self, recon_x, x, mu, log_var):
+        """Calculate VAE loss: reconstruction + KL divergence."""
+        # MSE reconstruction loss
+        recon_loss = F.mse_loss(recon_x, x, reduction='sum')
+        
+        # KL divergence
+        kl_loss = -0.5 * torch.sum(1 + log_var - mu.pow(2) - log_var.exp())
+        
+        return recon_loss + kl_loss
+
+class VAEOutputPreprocessor(BaseEstimator, TransformerMixin):
+    """
+    A custom transformer for sklearn Pipeline that uses a PyTorch Variational Autoencoder (VAE)
+    to perform dimensionality reduction on y (output) values during model training.
+    
+    Parameters
+    ----------
+    latent_dim : int, default=2
+        The dimensionality of the latent space.
+    
+    epochs : int, default=100
+        Number of epochs to train the VAE.
+    
+    batch_size : int, default=32
+        Batch size for training the VAE.
+    
[…]
