+++ docs/index.md
@@ -17,7 +17,7 @@
 - 🪄 **Low-Code**: Data-processing, model comparison, cross-validation, hyperparameter search and more in few lines of code
 - 🎯 **Optimized for Emulation**: Optimized for typical emulation scenarios with small to medium datasets (100s-1000s of points) with many inputs and outputs
 - 🔌 **Easy Integration**: All emulators are `scikit-learn` compatible, and the underlying `PyTorch` models can be extracted for custom use
-- 🔮 **Downstream Applications**: Still early days, but we've got prediction, sensitivity analysis, history matching and more
+- 🔮 **Downstream Applications**: Still early days, but we've got prediction, sensitivity analysis, calibration and more
 
 ## 🎓 State-of-the-Art Models
 
