+++ docs/_config.yml
@@ -46,6 +46,8 @@ sphinx:
   config:
     html_context:
       body_classes: home-page
+    html_theme_options:
+      theme: dark
     add_module_names: False
     autodoc_typehints: none
     autodoc_member_order: 'bysource'
