+++ _sources/reference/simulations/epidemic.rst
@@ -1,0 +1,6 @@
+autoemulate.simulations.epidemic
+==================================
+
+.. automodule:: autoemulate.simulations.epidemic
+   :members:
+   :show-inheritance:
