+++ _sources/community/faq/faq-users.md
@@ -1,0 +1,77 @@
+# First-Time Users' Frequently Asked Questions
+
+## General Questions
+
+1. What is `AutoEmulate`?
+   <!-- A brief description of what the package does, its main features, and its intended use case. -->
+   - A Python package that makes it easy to create emulators for complex simulations. It takes a set of simulation inputs `X` and outputs `y`, and automatically fits, optimises and evaluates various machine learning models to find the best emulator model. The emulator model can then be used as a drop-in replacement for the simulation, but will be much faster and computationally cheaper to evaluate. We have also implemented global sensitivity analysis as a common emulator application and working towards making `AutoEmulate` a true end-to-end package for building emulators.
+
+2. How do I know whether `AutoEmulate` is the right tool for me?
+   - You need to build an emulator for a simulation.
+   - You want to do global sensitivity analysis
+   - Your inputs `X` and outputs `y` are numeric and complete (we don't support missing data yet).
+   - You have one or more input parameters and one or more output variables.
+   - You have a small-ish dataset in the order of hundreds to few thousands of samples. All default emulator parameters and search spaces are optimised for smaller datasets.
+
+3. Does `AutoEmulate` support multi-output data?
+   - Yes, all models support multi-output data. Some do so natively, others are wrapped in a `MultiOutputRegressor`, which fits one model per target variable.
+
+4. Does `AutoEmulate` support temporal or spatial data?
+   - Not explicitly. The train-test split just takes a random subset as a test set, so does KFold cross-validation.
+
+5. `AutoEmulate` takes a long time to run on my dataset, why?
+   - The package fits a lot of models, in particular when hyperparameters are optimised. With say 8 default models and 5-fold cross-validation, this amounts to 40 model fits. With the addition of hyperparameter optimisation (n_iter=20), this results in 800 model fits. Some models such as Gaussian Processes and Neural Processes will take a long time to run on a CPU. However, don't despair! There is a [speeding up AutoEmulate guide](../../tutorials/02_speed.ipynb). As a rule of thumb, if your dataset is smaller than 1000 samples, you should be fine, if it's larger and you want to optimise hyperparameters, you might want to read the guide.
+
+## Usage Questions
+
+1. What data do I need to provide to `AutoEmulate` to build an emulator?
+   <!-- A simple example to get a new user started, possibly pointing to more detailed tutorials or documentation. -->
+   - You'll need two input objects: `X` and `y`. `X` is an ndarray / Pandas DataFrame of shape `(n_samples, n_parameters)` and `y` is an ndarray / Pandas DataFrame of shape `(n_samples, n_outputs)`. Each sample here is a simulation run, so each row of `X` corresponds to a set of input parameters and each row of `y` corresponds to the corresponding simulation output. You'll usually have created `X` using Latin hypercube sampling or similar methods, and `y` by running the simulation on these `X` inputs.
+
+2. Can I use `AutoEmulate` for commercial purposes?
+   <!-- Information on licensing and any restrictions on use. -->
+   - Yes. It's licensed under the MIT license, which allows for commercial use. See the [license](../../../LICENSE) for more information.
+
+## Advanced Usage
+
+1. Does AutoEmulate support parallel processing or high-performance computing (HPC) environments?
+   <!-- Details on the software's capabilities to leverage multi-threading, distributed computing, or HPC resources to speed up computations. -->
+   - Yes, [AutoEmulate.setup()](../../reference/compare.rst) has an `n_jobs` parameter which allows to parallelise cross-validation and hyperparameter optimisation. We are also working on GPU support for some models.
+
+2. Can AutoEmulate be integrated with other data analysis or simulation tools?
[…]
