### 2025-03-10 5cde5fc
Begin exploring refactor classes and examples
