### 2025-03-11 c52f60a
Add other base classes
