+++ autoemulate/refactor/autoemulate.py
@@ -1,0 +1,2 @@
+class AutoEmulateRefactor:
+    pass
