+++ autoemulate/refactor/__init__.py
@@ -1,0 +1,6 @@
+class AutoEmulate:
+    def __init__(self):
+        pass
+
+    def compare():
+        pass
