### 2025-03-11 550b62b
Revise base classes, add more examples
