+++ autoemulate/refactor/base.py
@@ -1,52 +1,210 @@
 import torch
-import numpy as np
+import torch.nn as nn
+from torch.optim.optimizer import Optimizer
+from sklearn.model_selection import KFold
+from torch.utils.data import DataLoader, Dataset
+
 from abc import ABC, abstractmethod
+from sklearn.base import BaseEstimator
+from sklearn.model_selection import cross_val_score
+import numpy as np
 
 from autoemulate.refactor.utils import convert_to_numpy, convert_to_tensor
 
 
-class BaseModel(ABC):
+class BaseEmulator(ABC):
     @abstractmethod
-    def fit(self, X: np.ndarray | torch.Tensor, y: np.ndarray | torch.Tensor):
+    def fit(
+        self,
+        dataloader: DataLoader,
+        criterion: nn.Module,
+        optimizer: Optimizer,
+        # TODO: refine with config/callbacks?
+        epochs=10,
+        device="cpu",
+    ):
         pass
 
     @abstractmethod
-    def predict(
-        self, X: np.ndarray | torch.Tensor
-    ) -> torch.Tensor | torch.distributions.Distribution:
+    def predict(self, dataloader: DataLoader, device: str = "cpu"):
         pass
 
-    # @abstractmethod
-    # def tune(self, X, y: torch.Tensor):
-    #     pass
+    @abstractmethod
+    def tune(
+        self,
+        dataset: Dataset,
+        criterion: nn.Module,
+        optimizer: Optimizer,
+        # TODO: could be replaced by general config
+        param_grid,
+        k: int = 3,
+        epochs: int = 10,
+        batch_size: int = 32,
+        device="cpu",
+    ):
+        pass
 
-    # @abstractmethod
-    # def run_cv(self, X: torch.Tensor, y: torch.Tensor):
-    #     pass
+    @abstractmethod
+    def cross_validate(
+        self,
+        dataset: Dataset,
+        criterion: nn.Module,
+        optimizer: Optimizer,
+        # TODO: could be replaced by general config
+        k=5,
+        epochs=10,
+        batch_size=32,
+        device="cpu",
+    ):
+        pass
 
 
-class PyTorchMixin(torch.nn.Module): ...
+class PyTorchBackend(nn.Module, BaseEmulator):
+    def __init__(self):
+        super().__init__()
+
+    def fit(self, dataloader, criterion, optimizer, epochs=10, device="cpu"):
+        self.to(device)
+        self.train()
+        for epoch in range(epochs):
+            total_loss = 0
+            for inputs, targets in dataloader:
+                inputs, targets = inputs.to(device), targets.to(device)
+                optimizer.zero_grad()
+                outputs = self(inputs)
+                loss = criterion(outputs, targets)
+                loss.backward()
+                optimizer.step()
+                total_loss += loss.item()
+            print(f"Epoch {epoch+1}/{epochs}, Loss: {total_loss / len(dataloader)}")
+
+    def predict(self, dataloader, device="cpu"):
+        self.to(device)
+        self.eval()
+        predictions = []
+        with torch.no_grad():
+            for inputs in dataloader:
+                inputs = inputs.to(device)
+                outputs = self(inputs)
+                predictions.append(outputs.cpu())
+        return torch.cat(predictions, dim=0)
+
+    def cross_validate(
+        self,
+        dataset,
+        criterion,
+        optimizer,
+        k=5,
+        epochs=10,
+        batch_size=32,
+        device="cpu",
+    ):
+        kf = KFold(n_splits=k, shuffle=True)
+        results = []
+        for fold, (train_idx, val_idx) in enumerate(kf.split(dataset)):
+            train_subset = torch.utils.data.Subset(dataset, train_idx)
[…]
