+++ sensitivity_analysis.py
@@ -1,304 +1,0 @@
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-from SALib.analyze.sobol import analyze
-from SALib.sample.sobol import sample
-
-from autoemulate.utils import _ensure_2d
-
-
-def _sensitivity_analysis(
-    model, problem=None, X=None, N=1024, conf_level=0.95, as_df=True
-):
-    """Perform Sobol sensitivity analysis on a fitted emulator.
-
-    Parameters:
-    -----------
-    model : fitted emulator model
-        The emulator model to analyze.
-    problem : dict
-        The problem definition, including 'num_vars', 'names', and 'bounds', optional 'output_names'.
-        Example:
-        ```python
-        problem = {
-            "num_vars": 2,
-            "names": ["x1", "x2"],
-            "bounds": [[0, 1], [0, 1]],
-        }
-        ```
-    N : int, optional
-        The number of samples to generate (default is 1024).
-    conf_level : float, optional
-        The confidence level for the confidence intervals (default is 0.95).
-    as_df : bool, optional
-        If True, return a pandas DataFrame (default is True).
-
-    Returns:
-    --------
-    pd.DataFrame or dict
-        If as_df is True, returns a long-format DataFrame with the sensitivity indices.
-        Otherwise, returns a dictionary where each key is the name of an output variable and each value is a dictionary
-        containing the Sobol indices keys ‘S1’, ‘S1_conf’, ‘ST’, and ‘ST_conf’, where each entry
-        is a list of length corresponding to the number of parameters.
-    """
-    Si = _sobol_analysis(model, problem, X, N, conf_level)
-
-    if as_df:
-        return _sobol_results_to_df(Si)
-    else:
-        return Si
-
-
-def _check_problem(problem):
-    """
-    Check that the problem definition is valid.
-    """
-    if not isinstance(problem, dict):
-        raise ValueError("problem must be a dictionary.")
-
-    if "num_vars" not in problem:
-        raise ValueError("problem must contain 'num_vars'.")
-    if "names" not in problem:
-        raise ValueError("problem must contain 'names'.")
-    if "bounds" not in problem:
-        raise ValueError("problem must contain 'bounds'.")
-
-    if len(problem["names"]) != problem["num_vars"]:
-        raise ValueError("Length of 'names' must match 'num_vars'.")
-    if len(problem["bounds"]) != problem["num_vars"]:
-        raise ValueError("Length of 'bounds' must match 'num_vars'.")
-
-    return problem
-
-
-def _get_output_names(problem, num_outputs):
-    """
-    Get the output names from the problem definition or generate default names.
-    """
-    # check if output_names is given
-    if "output_names" not in problem:
-        output_names = [f"y{i+1}" for i in range(num_outputs)]
-    else:
-        if isinstance(problem["output_names"], list):
-            output_names = problem["output_names"]
-        else:
-            raise ValueError("'output_names' must be a list of strings.")
-
-    return output_names
-
-
-def _generate_problem(X):
-    """
-    Generate a problem definition from a design matrix.
-    """
-    if X.ndim == 1:
-        raise ValueError("X must be a 2D array.")
-
-    return {
-        "num_vars": X.shape[1],
-        "names": [f"x{i+1}" for i in range(X.shape[1])],
-        "bounds": [[X[:, i].min(), X[:, i].max()] for i in range(X.shape[1])],
-    }
-
-
-def _sobol_analysis(model, problem=None, X=None, N=1024, conf_level=0.95):
-    """
-    Perform Sobol sensitivity analysis on a fitted emulator.
-
-    Parameters:
-    -----------
-    model : fitted emulator model
-        The emulator model to analyze.
-    problem : dict
[…]
