### 2025-03-12 756a7a1
remove changes not needed in context of new landing page
