+++ docs/_config.yml
@@ -42,12 +42,7 @@ sphinx:
     - 'sphinx.ext.autodoc'
     - 'sphinx.ext.napoleon'
     - 'sphinx.ext.viewcode'
-  extra_css: ["_static/homepage.css"]
   config:
-    html_context:
-      body_classes: home-page
-    html_theme_options:
-      theme: dark
     add_module_names: False
     autodoc_typehints: none
     autodoc_member_order: 'bysource'
