+++ docs/index.md
@@ -11,42 +11,8 @@
 🚀 Get started
 ```
 
-## ✨ Why AutoEmulate?
-
-- 🛠️ **Diverse Emulators**: From classic Radial Basis Functions to cutting-edge Neural Processes
-- 🪄 **Low-Code**: Data-processing, model comparison, cross-validation, hyperparameter search and more in few lines of code
-- 🎯 **Optimized for Emulation**: Optimized for typical emulation scenarios with small to medium datasets (100s-1000s of points) with many inputs and outputs
-- 🔌 **Easy Integration**: All emulators are `scikit-learn` compatible, and the underlying `PyTorch` models can be extracted for custom use
-- 🔮 **Downstream Applications**: Still early days, but we've got prediction, sensitivity analysis, history matching and more
-
-## 🎓 State-of-the-Art Models
-
-::::{grid} 1 1 2 3
-:gutter: 2
-
-:::{grid-item} 📈 **Classical**
-
-- Radial Basis Functions
-- Second Order Polynomials
-
-:::
-
-:::{grid-item} 🌳 **Machine Learning**
-
-- Random Forests
-- Gradient Boosting
-- Support Vector Machines
-- LightGBM
-
-:::
-
-:::{grid-item} <img src="https://pytorch.org/assets/images/pytorch-logo.png" height="16"/> **Deep Learning**
-
-- Multi-output / Multi-task Gaussian Processes
-- Conditional Neural Processes
-
-:::
-::::
+This website provides detailed documentation on how to use `AutoEmulate`, including installation instructions, tutorials, and a comprehensive API reference.
+For an overview of the project, checkout [autoemulate.com](https://autoemulate.com).
 
 ## 🔗 Quick Links
 
