### 2025-03-12 b8f4059
clarify instructions for building docs locally
