### 2025-03-13 badcf1f
add initial base class design for linear regression
Co-authored-by: Sam Greenbury <sgreenbury@users.noreply.github.com>
