+++ autoemulate/emulators/base.py
@@ -1,0 +1,190 @@
+from abc import ABC, abstractmethod
+from torch import nn
+from torch.utils.data import DataLoader, TensorDataset
+from torch import Tensor
+import torch.optim as optim
+import torch
+import torch.nn as nn
+import torch.optim as optim
+import numpy as np
+from torch.utils.data import DataLoader, TensorDataset, Dataset
+from typing import Union, Tuple, Optional
+from gpytorch.models import ExactGP
+import gpytorch
+from dataclasses import dataclass
+
+
+class BaseEmulator(ABC):
+    """The interface containing methods on emulators that are
+    expected by downstream dependents. This includes:
+    - `AutoEmulate`
+    - Active learning implementations
+    """
+    def __call__(self, *args, **kwds):
+        return self.predict(*args, **kwds)
+
+    @abstractmethod
+    def fit(self, X: DataLoader):
+        pass
+
+    @abstractmethod
+    def predict(self, X: DataLoader):
+        pass
+
+    @abstractmethod
+    def tune(self, X: Dataset):
+        pass
+
+    @abstractmethod
+    def cross_validate(self, X: Dataset):
+        pass
+
+
+# Maybe come back later
+# class CommonTorchBackend(nn.Module, BaseEmulator):
+#     pass
+
+class PyTorchBackend(nn.Module, BaseEmulator):
+    """PyTorchBackend is a torch model and implements the base class.
+    This provides default implementations to further subclasses.
+    This means that models can subclass and only need to implement
+    `.forward()` to have an emulator to be run in `AutoEmulate`"""
+
+    def fit(self,
+                   X: np.ndarray | torch.Tensor | DataLoader,
+                   y: np.ndarray | None | torch.Tensor = None,
+                   epochs: int = 3,
+                   batch_size: int = 1,
+                   verbose: bool = True) -> list:
+        """
+        Train the linear regression model.
+
+        Args:
+            X: Input features as numpy array, PyTorch tensor, or DataLoader
+            y: Target values (not needed if X is a DataLoader)
+            epochs: Number of training epochs
+            batch_size: Batch size (used only when X is not a DataLoader)
+            verbose: Whether to print progress
+
+        Returns:
+            List of loss values per epoch
+        """
+        self.train()  # Set model to training mode
+        loss_history = []
+
+        # Convert input to DataLoader if not already
+        if isinstance(X, np.ndarray):
+            X = torch.tensor(X, dtype=torch.float32)
+        if isinstance(y, np.ndarray):
+            y = torch.tensor(y, dtype=torch.float32)
+
+        if isinstance(X, (torch.Tensor, np.ndarray)) and y is not None:
+            dataset = TensorDataset(X, y)
+            dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True)
+        elif isinstance(X, DataLoader):
+            dataloader = X
+        else:
+            raise ValueError("Unsupported type for X. Must be numpy array, PyTorch tensor, or DataLoader.")
+
+
+        # Training loop
+        for epoch in range(epochs):
+            epoch_loss = 0.0
+            batches = 0
+
+            for X_batch, y_batch in dataloader:
+
+                # Forward pass
+                y_pred = self.forward(X_batch)
+                loss = self.loss_fn(y_pred, y_batch)
+
+                # Backward pass and optimize
+                self.optimizer.zero_grad()
+                loss.backward()
+                self.optimizer.step()
+
[…]
