### 2025-03-13 0e660d4
Midway through GPytorch implementation
Co-authored-by: Sam Greenbury <sgreenbury@users.noreply.github.com>
