+++ autoemulate/emulators/base.py
@@ -1,17 +1,12 @@
 from abc import ABC, abstractmethod
 from torch import nn
-from torch.utils.data import DataLoader, TensorDataset
+from torch.utils.data import DataLoader, TensorDataset, Dataset
 from torch import Tensor
 import torch.optim as optim
 import torch
-import torch.nn as nn
-import torch.optim as optim
 import numpy as np
-from torch.utils.data import DataLoader, TensorDataset, Dataset
-from typing import Union, Tuple, Optional
-from gpytorch.models import ExactGP
 import gpytorch
-from dataclasses import dataclass
+from gpytorch.likelihoods.likelihood import Likelihood
 
 
 class BaseEmulator(ABC):
@@ -20,6 +15,7 @@ class BaseEmulator(ABC):
     - `AutoEmulate`
     - Active learning implementations
     """
+
     def __call__(self, *args, **kwds):
         return self.predict(*args, **kwds)
 
@@ -40,22 +36,64 @@ class BaseEmulator(ABC):
         pass
 
 
+class InputTypeMixin:
+    def convert(
+        self,
+        X: np.ndarray | torch.Tensor | DataLoader,
+        y: np.ndarray | torch.Tensor | None = None,
+        batch_size: int = 16,
+        shuffle: bool = True,
+    ) -> DataLoader:
+        """
+        Mixin class to convert input data to pytorch DataLoaders.
+        """
+        # Convert input to DataLoader if not already
+        if isinstance(X, np.ndarray):
+            X = torch.tensor(X, dtype=torch.float32)
+        if isinstance(y, np.ndarray):
+            y = torch.tensor(y, dtype=torch.float32)
+
+        if isinstance(X, (torch.Tensor, np.ndarray)) and y is not None:
+            dataset = TensorDataset(X, y)
+            dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=shuffle)
+        elif isinstance(X, DataLoader) and y is None:
+            dataloader = X
+        else:
+            raise ValueError(
+                "Unsupported type for X. Must be numpy array, PyTorch tensor, or DataLoader."
+            )
+
+        return dataloader
+
+    # TODO: consider possible method for predict
+    # def convert_x(self, y: np.ndarray | torch.Tensor | Data) -> torch.Tensor:
+    #     if isinstance(y, np.ndarray):
+    #         y = torch.tensor(y, dtype=torch.float32)
+    #     else:
+    #         raise ValueError("Unsupported type for X. Must be numpy array, PyTorch tensor")
+    #     return y
+
+
 # Maybe come back later
 # class CommonTorchBackend(nn.Module, BaseEmulator):
 #     pass
 
-class PyTorchBackend(nn.Module, BaseEmulator):
+
+class PyTorchBackend(nn.Module, BaseEmulator, InputTypeMixin):
     """PyTorchBackend is a torch model and implements the base class.
     This provides default implementations to further subclasses.
     This means that models can subclass and only need to implement
     `.forward()` to have an emulator to be run in `AutoEmulate`"""
 
-    def fit(self,
-                   X: np.ndarray | torch.Tensor | DataLoader,
-                   y: np.ndarray | None | torch.Tensor = None,
-                   epochs: int = 3,
-                   batch_size: int = 1,
-                   verbose: bool = True) -> list:
+    def fit(
+        self,
+        X: np.ndarray | torch.Tensor | DataLoader,
+        y: np.ndarray | None | torch.Tensor = None,
+        epochs: int = 3,
+        batch_size: int = 1,
+        shuffle: bool = True,
+        verbose: bool = True,
[…]
