### 2025-03-13 73d33e3
Added reaction-diffusion simulator and tutorial
