### 2025-03-15 #7626 @crvernon
👋  @mastoffel - Your paper is looking good but please address the following:

- [ ] LINE 91:  The "g" in "gaussian" should be capitalized as well as "gpu".  You can achieve this by placing curly brackets around characters in your bib file that you wish to maintain formatting of.

We want to make sure the archive has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.

So here is what we have left to do on that front:

- [ ]  Conduct a GitHub release (if you haven't done so already) of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service.  Please ensure that the software archive uses the same license as the license you have posted on GitHub.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (make sure the list also matches your paper exactly).

- [ ]  Please respond with the DOI of the archived version here as well as the version of the software you released.
