### 2025-03-18 f286ad6
Fixed dimensionality reduction on outputs & embedded in scikit Pipeline
