### 2025-03-19 d5656a7
Added Turing pattern example (grey-scott diffusion model)
