### 2025-03-19 #7626 @mastoffel
Thanks @crvernon!

* Fixed the typos and updated the paper in the [joss branch](https://github.com/alan-turing-institute/autoemulate/tree/joss).
* The reviewed version is v0.2.1, and here is the [GitHub release](https://github.com/alan-turing-institute/autoemulate/releases/tag/v0.2.1).
* I archived the v0.2.1 release [on Zenodo](https://zenodo.org/records/15050266) with the same title, authors and licence.
* The Zenodo doi is [10.5281/zenodo.15050266](https://doi.org/10.5281/zenodo.15050266).
