### 2025-03-19 ecd8d54
Added autoencoder and variational autoencoders
