+++ autoemulate/preprocess_target.py
@@ -1,11 +1,12 @@
 import numpy as np
 import torch
 import torch.nn as nn
+import torch.optim as optim
 import torch.nn.functional as F
 from torch.utils.data import DataLoader, TensorDataset
 from sklearn.base import BaseEstimator, TransformerMixin
 
-class VAE(nn.Module):
+class VariationalAutoencoder(nn.Module):
     """
     Variational Autoencoder implementation in PyTorch.
     
@@ -18,8 +19,8 @@ class VAE(nn.Module):
     latent_dim : int
         Dimensionality of the latent space.
     """
-    def __init__(self, input_dim, hidden_dims, latent_dim):
-        super(VAE, self).__init__()
+    def __init__(self, input_dim, hidden_dims, latent_dim, verbose=False):
+        super(VariationalAutoencoder, self).__init__()
         
         # Encoder layers
         encoder_layers = []
@@ -69,259 +70,536 @@ class VAE(nn.Module):
         recon = self.decode(z)
         return recon, mu, log_var
 
-    def loss_function(self, recon_x, x, mu, log_var):
-        """Calculate VAE loss: reconstruction + KL divergence."""
+    def loss_function(self, recon_x, x, mu, log_var, beta=1.0):
+        """
+        Calculate VAE loss: reconstruction + KL divergence.
+        
+        Parameters
+        ----------
+        recon_x : torch.Tensor
+            Reconstructed input
+        x : torch.Tensor
+            Original input
+        mu : torch.Tensor
+            Mean of latent distribution
+        log_var : torch.Tensor
+            Log variance of latent distribution
+        beta : float, default=1.0
+            Weight for KL divergence term (beta-VAE)
+            
+        Returns
+        -------
+        total_loss : torch.Tensor
+            Combined reconstruction and KL divergence loss
+        """
         # MSE reconstruction loss
         recon_loss = F.mse_loss(recon_x, x, reduction='sum')
         
         # KL divergence
         kl_loss = -0.5 * torch.sum(1 + log_var - mu.pow(2) - log_var.exp())
         
-        return recon_loss + kl_loss
+        return recon_loss + beta * kl_loss
 
-class VAEOutputPreprocessor(BaseEstimator, TransformerMixin):
+class VariationalAutoencoderDimReducer(BaseEstimator, TransformerMixin):
     """
-    A custom transformer for sklearn Pipeline that uses a PyTorch Variational Autoencoder (VAE)
-    to perform dimensionality reduction on y (output) values during model training.
+    Sklearn-compatible wrapper for PyTorch Variational Autoencoder to use as dimensionality reducer.
+    Implements fit, transform, and inverse_transform methods required for sklearn pipelines.
     
     Parameters
     ----------
-    latent_dim : int, default=2
-        The dimensionality of the latent space.
-    
+    encoding_dim : int, default=10
+        Dimension of the encoded representation (latent space)
+    hidden_layers : list of int, default=None
+        List of hidden layer sizes for encoder
     epochs : int, default=100
-        Number of epochs to train the VAE.
-    
+        Number of training epochs
     batch_size : int, default=32
-        Batch size for training the VAE.
-    
+        Batch size for training
     learning_rate : float, default=0.001
-        Learning rate for the optimizer.
-    
-    hidden_dims : list, default=None
-        List of hidden dimensions for the encoder and decoder networks.
-        If None, [64, 32] will be used.
-    
-    verbose : bool, default=False
-        Whether to print training progress.
-        
+        Learning rate for optimizer
+    beta : float, default=1.0
+        Weight for KL divergence in the loss function (beta-VAE)
     device : str, default=None
-        Device to use for training ('cuda' or 'cpu'). If None, will use
[…]
