### 2025-03-20 1a63e98
Added various active learning classes. Work in progress.
