+++ autoemulate/learners/base.py
@@ -1,0 +1,333 @@
+# %%
+
+from dataclasses import dataclass, field, asdict, InitVar
+from abc import ABC, abstractmethod
+from typing import Union, Tuple, Dict, List, Any
+import numpy as np, torch
+from torcheval.metrics.functional import r2_score, mean_squared_error
+
+@dataclass(kw_only=True)
+class Base(ABC):
+
+    @staticmethod
+    def check_vector(X: torch.Tensor):
+        if not isinstance(X, torch.Tensor):
+            raise ValueError(f"Expected torch.Tensor, got {type(X)}")
+        elif X.ndim != 1:
+            raise ValueError(f"Expected 1D tensor, got {X.ndim}D")
+        else:
+            return X
+
+    @staticmethod
+    def check_matrix(X: torch.Tensor):
+        if not isinstance(X, torch.Tensor):
+            raise ValueError(f"Expected torch.Tensor, got {type(X)}")
+        elif X.ndim != 2:
+            raise ValueError(f"Expected 2D tensor, got {X.ndim}D")
+        else:
+            return X
+
+    @staticmethod
+    def check_pair(X: torch.Tensor, Y: torch.Tensor):
+        if X.shape[0] != Y.shape[0]:
+            raise ValueError("X and Y must have the same number of rows")
+        else:
+            return X, Y
+        
+    @staticmethod
+    def check_covariance(Y: torch.Tensor, Sigma: torch.Tensor):
+        # Full covariance matrix
+        if Sigma.shape == (Y.shape[0], Y.shape[1], Y.shape[1]):
+            return Sigma
+        # Diagonal covariance matrix
+        elif Sigma.shape == (Y.shape[0], Y.shape[1]):
+            return Sigma
+        # Scalar covariance matrix
+        elif Sigma.shape == (Y.shape[0],):
+            return Sigma
+        else:
+            raise ValueError("Invalid covariance matrix shape")
+        
+    @staticmethod
+    def trace(Sigma: torch.Tensor, d: int) -> torch.Tensor:
+        # A-optimal design criterion
+        if Sigma.dim() == 3 and Sigma.shape[1:] == (d, d):
+            return torch.diagonal(Sigma, dim1=1, dim2=2).sum(dim=1).mean()
+        elif Sigma.dim() == 2 and Sigma.shape[1] == d:
+            return Sigma.sum(dim=1).mean()
+        elif Sigma.dim() == 1:
+            return d * Sigma.mean()
+        else:
+            raise ValueError(f"Invalid covariance matrix shape: {Sigma.shape}")
+        
+    @staticmethod
+    def logdet(Sigma: torch.Tensor, dim: int):
+        # D-optimal design criterion
+        if len(Sigma.shape) == 3 and Sigma.shape[1:] == (dim, dim):
+            return torch.logdet(Sigma).mean()
+        elif len(Sigma.shape) == 2 and Sigma.shape[1] == dim:
+            return torch.sum(torch.log(Sigma), dim=1).mean()
+        elif len(Sigma.shape) == 1:
+            return dim * torch.log(Sigma).mean()
+        else:
+            raise ValueError(f"Invalid covariance matrix shape: {Sigma.shape}")
+        
+    @staticmethod
+    def max_eigval(Sigma: torch.Tensor) -> torch.Tensor:
+        # D-optimal design criterion
+        if Sigma.dim() == 3 and Sigma.shape[1:] == (Sigma.shape[1], Sigma.shape[1]):
+            eigvals = torch.linalg.eigvalsh(Sigma)
+            return eigvals[:, -1].mean() # Eigenvalues are sorted
+        elif Sigma.dim() == 2:
[…]
