### 2025-03-20 fcdbed3
Refined reaction-diffusion tutorial
