+++ autoemulate/preprocess_target.py
@@ -1,0 +1,300 @@
+import numpy as np
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import torch.optim as optim
+from sklearn.base import BaseEstimator
+from sklearn.base import clone
+from sklearn.base import TransformerMixin
+from sklearn.pipeline import Pipeline
+from torch.utils.data import DataLoader
+from torch.utils.data import TensorDataset
+
+
+class VAEOutputPreprocessor(BaseEstimator, TransformerMixin):
+    """
+    A scikit-learn compatible transformer that applies a Variational Autoencoder (VAE)
+    to target values (y) in a sklearn pipeline.
+
+    This class wraps a PyTorch VAE model to preprocess output data through the latent space
+    and provides inverse transformation capabilities.
+
+    Parameters
+    ----------
+    latent_dim : int, default=3
+        Dimension of the latent space.
+    hidden_dims : list, default=[64, 32]
+        Hidden dimensions for the VAE encoder and decoder networks.
+    epochs : int, default=100
+        Number of training epochs for the VAE.
+    batch_size : int, default=32
+        Batch size for training.
+    learning_rate : float, default=1e-3
+        Learning rate for the Adam optimizer.
+    device : str, default='cuda' if available else 'cpu'
+        Device to use for computation.
+    """
+
+    def __init__(
+        self,
+        latent_dim=3,
+        hidden_dims=None,
+        epochs=100,
+        batch_size=32,
+        learning_rate=1e-3,
+        device=None,
+        verbose=False
+    ):
+        self.latent_dim = latent_dim
+        self.hidden_dims = [64, 32] if hidden_dims is None else hidden_dims
+        self.epochs = epochs
+        self.batch_size = batch_size
+        self.learning_rate = learning_rate
+        self.device = device or ("cuda" if torch.cuda.is_available() else "cpu")
+        self.verbose = verbose 
+        self.vae = None
+        self.is_fitted_ = False
+
+    def _init_vae(self, input_dim):
+        """Initialize the VAE model with the appropriate input dimension."""
+
+        self.vae = VAE(
+            input_dim=input_dim,
+            hidden_dims=self.hidden_dims,
+            latent_dim=self.latent_dim,
+        ).to(self.device)
+
+    def _create_data_loader(self, y):
+        """Create a PyTorch DataLoader for the target values."""
+        y_tensor = torch.FloatTensor(y).to(self.device)
+        dataset = torch.utils.data.TensorDataset(y_tensor)
+        return torch.utils.data.DataLoader(
+            dataset, batch_size=self.batch_size, shuffle=True
+        )
+
+    def fit(self, X, y=None, **fit_params):
+        """
+        Fit the VAE model to the target values.
+
+        Parameters
+        ----------
+        X : array-like, shape (n_samples, n_features)
+            Input data (not used but required by the sklearn API).
+        y : array-like, shape (n_samples, n_outputs)
+            Target values to transform.
+
+        Returns
+        -------
+        self : object
+            Returns self.
+        """
+        if y is None:
+            raise ValueError(
+                "Target values (y) are required for VAEOutputPreprocessor."
+            )
+
+        # Ensure y is 2D
+        if y.ndim == 1:
+            y = y.reshape(-1, 1)
+
+        # Initialize VAE with input dimension from y
+        self.n_features_out_ = y.shape[1]
+        self._init_vae(input_dim=self.n_features_out_)
+
+        # Create data loader
+        data_loader = self._create_data_loader(y)
+
+        # Train the VAE
+        optimizer = optim.Adam(self.vae.parameters(), lr=self.learning_rate)
+        self.vae.train()
+
+        for epoch in range(self.epochs):
[…]
