### 2025-03-21 d391596
Added get_model function
