+++ autoemulate/preprocess_target.py
@@ -9,7 +9,45 @@ from sklearn.base import TransformerMixin
 from sklearn.pipeline import Pipeline
 from torch.utils.data import DataLoader
 from torch.utils.data import TensorDataset
+from sklearn.decomposition import PCA
 
+def get_dim_reducer(name, n_components=8, encoding_dim=8, hidden_layers=None, 
+                   epochs=1200, batch_size=32, beta=1.0, verbose=False):
+    """
+    Factory function to get a dimensionality reducer based on name.
+    
+    Parameters
+    ----------
+    name : str or None
+        Name of the dimensionality reducer to use.
+        Options:
+        - 'PCA': Principal Component Analysis
+        - 'AE': Autoencoder
+        - 'VAE': Variational Autoencoder
+        - None: No dimensionality reduction (returns None)
+        
+    Returns
+    -------
+    dim_reducer : object or None
+        Scikit-learn compatible dimensionality reducer or None if name is None.
+    """
+    if name is None:
+        return None
+    
+    # Return the appropriate dimensionality reducer
+    if name == 'PCA': 
+        return PCA(n_components=8)
+
+    elif name == 'VAE':
+        return VAEOutputPreprocessor(
+            encoding_dim=8,
+            hidden_layers=[64, 32],
+            epochs=1200,
+            batch_size=32,
+            verbose=False
+        )
+    else:
+        raise ValueError(f"Unknown dimensionality reducer: {name}")
 
 class VAEOutputPreprocessor(BaseEstimator, TransformerMixin):
     """
