### 2025-03-21 df31adc
Modified pipeline and add output dimensionality reducer
