### 2025-03-21 1a4dbe1
Fixed issues compare.py
