### 2025-03-21 fedfbbf
Fix issues in utils.py
