+++ autoemulate/utils.py
@@ -12,6 +12,7 @@ from sklearn.exceptions import ConvergenceWarning
 from sklearn.model_selection import KFold
 from sklearn.multioutput import MultiOutputRegressor
 from sklearn.pipeline import Pipeline
+from sklearn.compose import TransformedTargetRegressor
 
 
 # manage warnings -------------------------------------------------------------
@@ -94,6 +95,8 @@ def get_model_name(model):
         # If the model step is a MultiOutputRegressor, get the estimator
         if isinstance(step, MultiOutputRegressor):
             return step.estimator.model_name
+        elif isinstance(step, TransformedTargetRegressor):
+            return get_model_name(step.regressor)  # Unwrap TransformedTargetRegressor
         else:
             return step.model_name
 
@@ -101,6 +104,10 @@ def get_model_name(model):
     elif isinstance(model, MultiOutputRegressor):
         return model.estimator.model_name
 
+    # If the model is a TransformedTargetRegressor, unwrap it
+    elif isinstance(model, TransformedTargetRegressor):
+        return get_model_name(model.regressor)
+    
     # Otherwise, it's a standalone model
     else:
         return model.model_name
