### 2025-03-21 21142b8
Add reaction-diffusion tutorial
