+++ autoemulate/printing.py
@@ -51,13 +51,15 @@ def _print_setup(cls):
             str(cls.param_search_iters),
             str(cls.reduce_dim),
             str(cls.reduce_dim_output),
-            str(cls.dim_reducer
+            str(
+                cls.dim_reducer
                 if isinstance(cls.dim_reducer, str)
                 else cls.dim_reducer.__class__.__name__
                 if cls.dim_reducer is not None
                 else "None"
             ),
-            str(cls.dim_reducer_output
+            str(
+                cls.dim_reducer_output
                 if isinstance(cls.dim_reducer_output, str)
                 else cls.dim_reducer_output.__class__.__name__
                 if cls.dim_reducer_output is not None
@@ -100,7 +102,9 @@ def _print_setup(cls):
 
     # if cls.reduce_dim == False, remove the dim_reducer row
     if not cls.reduce_dim:
-        settings = settings.drop(["Dimensionality input reduction method (dim_reducer)"])
+        settings = settings.drop(
+            ["Dimensionality input reduction method (dim_reducer)"]
+        )
 
     # if cls.scale == False, remove the scaler row
     if not cls.scale:
