### 2025-03-22 #7626 @crvernon
@mastoffel - The "g" in "gaussian" still needs to be capitalized in LINE 91.  No need to conduct a new release after this fix.
