+++ paper/paper.bib
@@ -170,7 +170,7 @@
 }
 
 @article{gardner_gpytorch_2018,
-	title = {Gpytorch: {Blackbox} matrix-matrix {Gaussian} process inference with GPU acceleration},
+	title = {Gpytorch: {Blackbox} matrix-matrix {Gaussian} process inference with {GPU} acceleration},
 	volume = {31},
 	shorttitle = {Gpytorch},
 	url = {https://proceedings.neurips.cc/paper/2018/hash/27e8e17134dd7083b050476733207ea1-Abstract.html},
