### 2025-03-24 55c41c4
Added VAE
