### 2025-03-24 #7626 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
