+++ autoemulate/learners/base.py
@@ -1,16 +1,40 @@
 # %%
-
-from dataclasses import dataclass, field, asdict, InitVar
+from dataclasses import dataclass, field, InitVar
 from abc import ABC, abstractmethod
 from typing import Union, Tuple, Dict, List, Any
 import numpy as np, torch
-from torcheval.metrics.functional import r2_score, mean_squared_error
+from torcheval.metrics import MeanSquaredError, R2Score
+from tqdm import tqdm
+
 
 @dataclass(kw_only=True)
 class Base(ABC):
+    """
+    Base class for active learning simulation and emulation.
+
+    Provides utility methods for tensor validation and design criteria computations.
+    """
 
     @staticmethod
-    def check_vector(X: torch.Tensor):
+    def check_vector(X: torch.Tensor) -> torch.Tensor:
+        """
+        Validate that the input is a 1D torch.Tensor.
+
+        Parameters
+        ----------
+        X : torch.Tensor
+            Input tensor to validate.
+
+        Returns
+        -------
+        torch.Tensor
+            Validated 1D tensor.
+
+        Raises
+        ------
+        ValueError
+            If X is not a torch.Tensor or is not 1-dimensional.
+        """
         if not isinstance(X, torch.Tensor):
             raise ValueError(f"Expected torch.Tensor, got {type(X)}")
         elif X.ndim != 1:
@@ -19,7 +43,25 @@ class Base(ABC):
             return X
 
     @staticmethod
-    def check_matrix(X: torch.Tensor):
+    def check_matrix(X: torch.Tensor) -> torch.Tensor:
+        """
+        Validate that the input is a 2D torch.Tensor.
+
+        Parameters
+        ----------
+        X : torch.Tensor
+            Input tensor to validate.
+
+        Returns
+        -------
+        torch.Tensor
+            Validated 2D tensor.
+
+        Raises
+        ------
+        ValueError
+            If X is not a torch.Tensor or is not 2-dimensional.
+        """
         if not isinstance(X, torch.Tensor):
             raise ValueError(f"Expected torch.Tensor, got {type(X)}")
         elif X.ndim != 2:
@@ -28,29 +70,85 @@ class Base(ABC):
             return X
 
     @staticmethod
-    def check_pair(X: torch.Tensor, Y: torch.Tensor):
+    def check_pair(X: torch.Tensor, Y: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+        """
+        Validate that two tensors have the same number of rows.
+
+        Parameters
+        ----------
+        X : torch.Tensor
+            First tensor.
+        Y : torch.Tensor
+            Second tensor.
+
+        Returns
+        -------
+        Tuple[torch.Tensor, torch.Tensor]
+            The validated pair of tensors.
+
+        Raises
+        ------
+        ValueError
+            If X and Y do not have the same number of rows.
+        """
         if X.shape[0] != Y.shape[0]:
             raise ValueError("X and Y must have the same number of rows")
         else:
             return X, Y
-        
+
     @staticmethod
-    def check_covariance(Y: torch.Tensor, Sigma: torch.Tensor):
-        # Full covariance matrix
+    def check_covariance(Y: torch.Tensor, Sigma: torch.Tensor) -> torch.Tensor:
+        """
+        Validate and return the covariance matrix.
+
+        Parameters
+        ----------
+        Y : torch.Tensor
+            Output tensor.
+        Sigma : torch.Tensor
+            Covariance matrix, which may be full, diagonal, or a scalar per sample.
+
+        Returns
+        -------
+        torch.Tensor
+            Validated covariance matrix.
+
+        Raises
+        ------
+        ValueError
+            If Sigma does not have a valid shape relative to Y.
+        """
[…]
