### 2025-03-24 6d5dd3b
This is a large commit - I have implemented an outer loop around the loop which runs through all models and select the best model. This outerloop goes through various pre-processing methods of data. Plenty of work is still required, name of variables should become consistent with the code base, input and output processes should become integrated. However for now , the whole pipeline and plotting functionalities run with no error
