+++ autoemulate/model_processing.py
@@ -33,18 +33,8 @@ def _turn_models_into_multioutput(models, y):
     return models_multi
 
 
-def _wrap_models_in_pipeline(
-    models,
-    scale,
-    scaler,
-    reduce_dim,
-    dim_reducer,
-    scale_output=False,
-    reduce_dim_output=False,
-    scaler_output=None,
-    dim_reducer_output=None,
-):
-    """Wrap models in a pipeline.
+def _wrap_models_in_pipeline(models, scale, scaler, reduce_dim, dim_reducer):
+    """Wrap models in a pipeline if scale is True.
 
     Parameters
     ----------
@@ -56,51 +46,27 @@ def _wrap_models_in_pipeline(
         Scaler to use.
     reduce_dim : bool
         Whether to reduce the dimensionality of the data.
-    dim_reducer : sklearn.decomposition object or str
-        Dimensionality reduction method to use or its name.
-    scale_output : bool, default=False
-        Whether to scale the output data.
-    reduce_dim_output : bool, default=False
-        Whether to reduce the dimensionality of the output data.
-    scaler_output : sklearn.preprocessing object
-        Scaler to use for outputs. If None and scale_output is True, a StandardScaler will be used.
-    dim_reducer_output : sklearn.decomposition object or str
-        Dimensionality reduction method to use for outputs or its name.
-        If None and reduce_dim_output is True, a PCA will be used.  #TODO
+    dim_reducer : sklearn.decomposition object
+        Dimensionality reduction method to use.
 
     Returns
     -------
     models_scaled : dict
         dict of model_names: model instances, with scaled models wrapped in a pipeline.
     """
+
     models_piped = []
+
     for model in models:
         steps = []
-        if reduce_dim_output:
-            # Only call get_dim_reducer if dim_reducer is a string
-            reducer = get_dim_reducer(name=dim_reducer_output)
-            steps.append(("Dimentionality reducer for output ", reducer))
-
-        # Add X preprocessing steps
         if scale:
             steps.append(("scaler", scaler))
         if reduce_dim:
             steps.append(("dim_reducer", dim_reducer))
-
-        # Add the model as the final step
         steps.append(("model", model))
+        # without scaling or dim reduction, the model is the only step
+        models_piped.append(Pipeline(steps))
 
-        # Use YAwarePipeline if we need to transform y, otherwise use standard Pipeline
-        if reduce_dim_output:
-            pipeline = AutoEmulatePipeline(steps)
-        else:
-            pipeline = Pipeline(steps)
-
-        # Explicitly ensure model_name is set
-        if hasattr(model, "model_name"):
-            pipeline.model_name = model.model_name
-
-        models_piped.append(pipeline)
     return models_piped
 
 
@@ -148,115 +114,6 @@ def _process_models(
     models = model_registry.get_models(model_names)
     models_multi = _turn_models_into_multioutput(models, y)
     models_scaled = _wrap_models_in_pipeline(
-        models_multi,
-        scale,
-        scaler,
-        reduce_dim,
-        dim_reducer,
-        scale_output,
-        reduce_dim_output,
-        scaler_output,
-        dim_reducer_output,
+        models_multi, scale, scaler, reduce_dim, dim_reducer
     )
     return models_scaled
-
-
-class AutoEmulatePipeline(Pipeline):
-    """Pipeline that can modify both X and y, while preserving all Pipeline attributes."""
-
-    def __init__(self, steps, memory=None, verbose=True):
-        # Initialize with parent class - this will set up all the Pipeline internals
-        super().__init__(steps=steps, memory=memory, verbose=verbose)
-
-        # Track the steps separately to avoid property issues
-        self._y_transformers = []
-        self._x_steps = []
-
-        # Process the steps after super().__init__() has been called
[…]
