+++ autoemulate/compare.py
@@ -1,3 +1,5 @@
+import copy
+
 import numpy as np
 import pandas as pd
 from sklearn.base import BaseEstimator
@@ -18,6 +20,7 @@ from autoemulate.metrics import METRIC_REGISTRY
 from autoemulate.model_processing import _process_models
 from autoemulate.plotting import _plot_cv
 from autoemulate.plotting import _plot_model
+from autoemulate.preprocess_target import get_dim_reducer
 from autoemulate.printing import _print_setup
 from autoemulate.save import ModelSerialiser
 from autoemulate.sensitivity_analysis import _plot_sensitivity_analysis
@@ -59,6 +62,7 @@ class AutoEmulate:
         scaler_output=StandardScaler(),
         reduce_dim_output=False,
         dim_reducer_output=PCA(),
+        preprocessing_methods=None,
         cross_validator=KFold(
             n_splits=5, shuffle=True, random_state=np.random.randint(1e5)
         ),
@@ -157,6 +161,7 @@ class AutoEmulate:
         self.dim_reducer_output = dim_reducer_output
         self.reduce_dim_output = reduce_dim_output
         self.cv_results = {}
+        self.preprocessing_methods = preprocessing_methods
 
         if print_setup:
             self.print_setup()
@@ -212,11 +217,12 @@ class AutoEmulate:
 
         # Freshly initialise scores dataframe when running compare()
         self.scores_df = pd.DataFrame(
-            columns=["model", "short", "metric", "fold", "score"]
+            columns=["model", "short", "preprocessing", "metric", "fold", "score"]
         ).astype(
             {
                 "model": "object",
                 "short": "object",
+                "preprocessing": "object",
                 "metric": "object",
                 "fold": "int64",
                 "score": "float64",
@@ -228,97 +234,349 @@ class AutoEmulate:
         else:
             pb_text = "Cross-validating"
 
-        with tqdm(total=len(self.models), desc="Initializing") as pbar:
-            for i, model in enumerate(self.models):
-                model_name = get_model_name(model)
-                pbar.set_description(f"{pb_text} {model_name}")
-                with _redirect_warnings(self.logger):
-                    try:
-                        # hyperparameter search
-                        if self.param_search:
-                            self.models[i] = _optimize_params(
+        # Initialize dictionary to store results
+        self.preprocessing_results = {}
+
+        # Handle case when no preprocessing methods are specified
+        if self.preprocessing_methods is None or len(self.preprocessing_methods) == 0:
+            # Default to no preprocessing
+            prep_name = "None"
+            self.preprocessing_results[prep_name] = {
+                "models": copy.deepcopy(self.models),
+                "cv_results": {},
+                "best_model": None,
+                "transformer": None,
+                "params": {},
+            }
+
+            with tqdm(total=len(self.models), desc=pb_text) as pbar:
+                for i, model in enumerate(
+                    self.preprocessing_results[prep_name]["models"]
+                ):
+                    model_name = get_model_name(model)
+                    pbar.set_description(f"{pb_text} {model_name}")
+
+                    with _redirect_warnings(self.logger):
+                        try:
+                            # hyperparameter search
+                            if self.param_search:
+                                self.preprocessing_results[prep_name]["models"][
+                                    i
+                                ] = _optimize_params(
+                                    X=self.X[self.train_idxs],
+                                    y=self.y[self.train_idxs],
+                                    cv=self.cross_validator,
+                                    model=model,
+                                    search_type=self.search_type,
+                                    niter=self.param_search_iters,
+                                    param_space=None,
+                                    n_jobs=self.n_jobs,
+                                    logger=self.logger,
+                                )
+
+                            # run cross validation
+                            fitted_model, cv_results = _run_cv(
                                 X=self.X[self.train_idxs],
[…]

[…]
