### 2025-03-24 aee95a4
Added registry functionality to learners
